+++ dask_geomodeling/tests/test_geometry.py
@@ -1357,6 +1357,7 @@ class TestFieldOperations(unittest.TestCase):
                 "col_source": float(i * 2 + 1),
                 "col_choice_1": chr(i + 65),  # 'A'
                 "col_choice_2": chr(i + 70),  # 'F'
+                "none": None,
             }
             for i, x in enumerate(values)
         ]
@@ -1477,7 +1478,15 @@ class TestFieldOperations(unittest.TestCase):
         self.assertEqual("B", values[2])  # 1.2
         self.assertEqual("C", values[3])  # 5.
         self.assertEqual("C", values[4])  # inf
-        self.assertTrue(np.isnan(values[5]))  # nan
+        self.assertTrue(np.isnan(values[5]))  # nan.
+
+    def test_classify_none(self):
+        series = field_operations.Classify(
+            self.source["none"], bins=[0, 0.2], labels=["A"]
+        )
+        result = series.get_data(**self.request)
+        values = result.values
+        self.assertTrue(all([np.isnan(x) for x in values]))
 
     def test_classify_from_columns_empty(self):
         view = field_operations.ClassifyFromColumns(
@@ -1487,6 +1496,14 @@ class TestFieldOperations(unittest.TestCase):
             mode="intersects", projection="EPSG:3857", geometry=box(0, 0, 0, 0)
         )
         self.assertEqual(0, len(result))
+
+    def test_classify_from_columns_none(self):
+        series = field_operations.ClassifyFromColumns(
+            self.source, "none", ["id_value"], labels=["A", "B"]
+        )
+        result = series.get_data(**self.request)
+        values = result.values
+        self.assertTrue(all([np.isnan(x) for x in values]))
 
     def test_classify_from_columns_varying_bin(self):
         series = field_operations.ClassifyFromColumns(