+++ CHANGES.rst
@@ -8,6 +8,11 @@ Changelog of dask-geomodeling
 - Allow up to 1E-7 in the GeoTransform 'tilt' terms to account for possible
   float32 imprecision.
 
+- Handle Nones in geometry.field_operations.Classify and ClassifyFromColumns.
+
+- Validate if labels are unique in geometry.field_operations.Classify and
+  ClassifyFromColumns.
+
 
 2.2.3 (2020-02-28)
 ------------------